+++ plugins/imprint/engine/imprint/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 
 from .theme import Theme
 from .render import build